compute_space_cli/pyproject.toml
[project]
name = "oh"
version = "0.1.0"
description = "OpenHost CLI — manage apps on your compute space"
requires-python = ">=3.12"
dependencies = [
    "attrs",
    "cattrs",
    "cappa",
    "httpx",
    "tomli-w",
]

[project.scripts]
oh = "compute_space_cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["compute_space_cli"]
